Clarify mode 5 fixed point conversion
Agent-Logs-Url: https://github.com/rmstdope/neser/sessions/f19d8806-63f4-4e53-b6b2-f6a4e3413cb3

+++ src/gba/ppu/mod.rs
@@ -1020,8 +1020,8 @@ impl Ppu {
                     continue;
                 }
 
-                let px = (sample_x as u32 >> 8) as usize;
-                let py = (sample_y as u32 >> 8) as usize;
+                let px = (sample_x >> 8) as usize;
+                let py = (sample_y >> 8) as usize;
                 if px >= MODE5_WIDTH || py >= MODE5_HEIGHT {
                     continue;
                 }
